[IRGen] NFC, move MetadataSource into its own header
